moving io to more classes
